feat: Apartments, Services and Sponsors Seeder

+++ database/migrations/2024_05_02_144048_create_sponsors_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
         Schema::create('sponsors', function (Blueprint $table) {
             $table->id();
             $table->string('name', 40);
-            $table->tinyInteger('duration');
+            $table->smallInteger('duration');
             $table->decimal('price', 4, 2, true);
             $table->timestamps();
         });
